+++ src/wrappers/React.js
@@ -24,12 +24,19 @@ const makeReactContainer = Component => {
     }
 
     render () {
-      const { children, ...rest } = this.state
+      const {
+        children,
+        // Vue attaches an event handler, but it is missing an event name, so
+        // it ends up using an empty string. Prevent passing an empty string
+        // named prop to React.
+        '': _invoker,
+        ...rest
+      } = this.state
       const wrappedChildren = this.wrapVueChildren(children)
 
       return (
         <Component {...rest}>
-          <VueWrapper component={wrappedChildren} />
+          {children && <VueWrapper component={wrappedChildren} />}
         </Component>
       )
     }
@@ -44,12 +51,13 @@ export default {
   methods: {
     mountReactComponent (component) {
       const Component = makeReactContainer(component)
+      const children = this.$slots.default !== undefined ? { children: this.$slots.default } : {}
       ReactDOM.render(
         <Component
           {...this.$props.passedProps}
           {...this.$attrs}
           {...this.$listeners}
-          children={this.$slots.default}
+          {...children}
           ref={ref => (this.reactComponentRef = ref)}
         />,
         this.$refs.react
@@ -67,7 +75,11 @@ export default {
      * AFAIK, this is the only way to update children. It doesn't seem to be possible to watch
      * `$slots` or `$children`.
      */
-    this.reactComponentRef.setState({ children: this.$slots.default })
+    if (this.$slots.default !== undefined) {
+      this.reactComponentRef.setState({ children: this.$slots.default })
+    } else {
+      this.reactComponentRef.setState({ children: null })
+    }
   },
   inheritAttrs: false,
   watch: {