+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Fixed
+- Only pass wrapped children if children exist to pass prop type validation. (#27)
+
 ## [0.2.0] - 2017-11-10
 ### Updated
 - Rollup 0.51